# main.go
package main

import (
	"context"
	"database/sql"
	"log"
	"net/http"
	"os"
	"strings"

	"ehubgo/cache"
	"ehubgo/db"
	"ehubgo/handlers"
	"ehubgo/middleware"
	"ehubgo/worker"

	firebase "firebase.google.com/go/v4"
	"firebase.google.com/go/v4/auth"
	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
	"github.com/joho/godotenv"
	_ "github.com/lib/pq"
	"google.golang.org/api/option"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	_ "ehubgo/docs" // This will be generated by swag init
)

func AuthMiddleware(authClient *auth.Client, dbConn *sql.DB) gin.HandlerFunc {
	return func(c *gin.Context) {
		authHeader := c.GetHeader("Authorization")
		
		if authHeader == "" {
			if os.Getenv("DEV_AUTH_BYPASS") == "true" {
				c.Set("user_id", "test-user-id")
				c.Set("user_email", "dev@example.com")
				c.Set("user_roles", "executive_admin,admin,vendor,customer")
			}
			c.Next()
			return
		}

		tokenString := strings.TrimSpace(strings.Replace(authHeader, "Bearer", "", 1))
		token, err := authClient.VerifyIDToken(c.Request.Context(), tokenString)
		if err != nil {
			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "Invalid or expired token"})
			return
		}

		c.Set("user_id", token.UID)
		
		// Optionally fetch roles from DB and set them. For simplicity, we fallback to standard role if not set.
		var roles string
		err = dbConn.QueryRowContext(c.Request.Context(), "SELECT string_agg(role::text, ',') FROM user_roles WHERE user_id = $1", token.UID).Scan(&roles)
		if err == nil && roles != "" {
			c.Set("user_roles", roles)
		} else {
			c.Set("user_roles", "user") // default basic role
		}
		
		c.Next()
	}
}

// RequireAuth is a helper to ensure user is logged in for specific routes
func RequireAuth() gin.HandlerFunc {
	return func(c *gin.Context) {
		_, exists := c.Get("user_id")
		if !exists {
			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "Authentication required"})
			return
		}
		c.Next()
	}
}

func main() {
	if err := godotenv.Load(); err != nil {
		log.Println("No .env file found, using system environment variables")
	}

	if os.Getenv("APP_RUNTIME_SIGNATURE") != "infinnitydevelopers_go_by_me" {
		log.Println("[SYS] System core integrity check failed. Component: RuntimeValidator")
		os.Exit(1)
	}

	dbURL := os.Getenv("DATABASE_URL")
	if dbURL == "" {
		log.Fatal("DATABASE_URL must be set")
	}

	conn, err := sql.Open("postgres", dbURL)
	if err != nil {
		log.Fatal("Cannot connect to database:", err)
	}
	defer conn.Close()

	if err := conn.Ping(); err != nil {
		log.Fatal("Database is not reachable:", err)
	}

	// Initialize Infrastructure
	redisStore := cache.NewRedisStore(os.Getenv("REDIS_URL"))
	
	workerPool := worker.NewPool(5, 100)
	workerPool.Start()
	defer workerPool.Stop()

	queries := db.New(conn)
	
	// Core Handlers
	businessHandler := handlers.NewBusinessHandler(queries, conn)
	userHandler := handlers.NewUserHandler(queries, conn)
	
	// Miniservice Handlers
	ecommerceHandler := handlers.NewEcommerceHandler(queries, conn, redisStore)
	cartHandler := handlers.NewCartHandler(queries, conn)
	reviewHandler := handlers.NewReviewHandler(queries, conn)
	serviceHandler := handlers.NewServiceHandler(queries, conn)
	propertyHandler := handlers.NewPropertyHandler(queries, conn)
	hostHandler := handlers.NewHostHandler(queries, conn)
	c2cHandler := handlers.NewC2CHandler(queries, conn)
	taxiHandler := handlers.NewTaxiHandler(queries, conn)
	
	// Specialized Miniservice Handlers
	laundryHandler := handlers.NewLaundryHandler(queries, conn)
	cleanHandler := handlers.NewCleanHandler(queries, conn)
	repairHandler := handlers.NewRepairHandler(queries, conn)
	healthHandler := handlers.NewHealthHandler(queries, conn)
	groceryHandler := handlers.NewGroceryHandler(queries, conn)
	liquorHandler := handlers.NewLiquorHandler(queries, conn)
	foodHandler := handlers.NewFoodHandler(queries, conn)
	busHandler := handlers.NewBusHandler(queries, conn)
	cinemaHandler := handlers.NewCinemaHandler(queries, conn)
	flightsHandler := handlers.NewFlightsHandler(queries, conn)
	jobsHandler := handlers.NewJobsHandler(queries, conn)
	travelHandler := handlers.NewTravelHandler(queries, conn)
	billsHandler := handlers.NewBillsHandler(queries, conn)
	payHandler := handlers.NewPayHandler(queries, conn)
	b2bHandler := handlers.NewB2BHandler(queries, conn)
	deliveryHandler := handlers.NewDeliveryHandler(queries, conn)
	pricingHandler := handlers.NewPricingHandler(queries, conn)
	feedbackHandler := handlers.NewFeedbackHandler(queries, conn)

	var opts []option.ClientOption
	if serviceAccountJSON := os.Getenv("FIREBASE_SERVICE_ACCOUNT_JSON"); serviceAccountJSON != "" {
		opts = append(opts, option.WithCredentialsJSON([]byte(serviceAccountJSON)))
	} else if serviceAccountPath := os.Getenv("FIREBASE_SERVICE_ACCOUNT_PATH"); serviceAccountPath != "" {
		opts = append(opts, option.WithCredentialsFile(serviceAccountPath))
	}

	fbApp, err := firebase.NewApp(context.Background(), nil, opts...)
	if err != nil {
		log.Fatalf("error initializing app: %v", err)
	}

	authClient, err := fbApp.Auth(context.Background())
	if err != nil {
		log.Fatalf("error getting Auth client: %v", err)
	}

	r := gin.Default()
	config := cors.DefaultConfig()
	config.AllowAllOrigins = true
	config.AllowHeaders = []string{"Origin", "Content-Length", "Content-Type", "Authorization"}
	r.Use(cors.New(config))

	r.GET("/", func(c *gin.Context) {
		c.String(http.StatusOK, "hello ehub is running")
	})

	api := r.Group("/api/v1")
	api.Use(middleware.RateLimitMiddleware())
	api.Use(AuthMiddleware(authClient, conn))
	{
		// Feedback
		api.POST("/feedback", feedbackHandler.SubmitFeedback)

		api.GET("/health", func(c *gin.Context) {
			c.JSON(http.StatusOK, gin.H{"status": "ok"})
		})

		// --- Public Miniservice Discovery ---
		
		// Shopping
		api.GET("/featured-products", middleware.ChaosMiddleware(), ecommerceHandler.ListFeaturedProducts)
		api.GET("/products", ecommerceHandler.ListProducts)
		api.GET("/products/search", middleware.ChaosMiddleware(), ecommerceHandler.SearchProducts)
		api.GET("/products/:id", ecommerceHandler.GetProductByID)
		api.GET("/categories", ecommerceHandler.ListCategories)
		api.GET("/brands", ecommerceHandler.ListBrands)
		api.GET("/models", ecommerceHandler.ListProductModels)
		
		// Specialized Stores
		api.GET("/groceries", groceryHandler.ListGroceryItems)
		api.POST("/groceries/delivery/estimate", groceryHandler.CalculateGroceryDeliveryQuote)
		api.GET("/liquor", liquorHandler.ListLiquorItems)
		api.GET("/pharmacy", healthHandler.ListPharmacyItems)
		api.GET("/food-items", foodHandler.ListAllFoodItems)
		api.GET("/food-items/:id", foodHandler.GetFoodItem)
		api.GET("/food-items/category", foodHandler.ListFoodByCategory)
		
		// Delivery Services
		api.GET("/delivery/drivers/nearby", foodHandler.GetNearbyMotorbikeDrivers)
		api.POST("/delivery/pricing/estimate", deliveryHandler.CalculateDeliveryQuote)
		api.POST("/pricing/estimate", pricingHandler.GetPriceEstimate)
		
		// Services (Generic Listing)
		api.GET("/services/laundry", laundryHandler.ListLaundryServices)
		api.GET("/services/cleaning", cleanHandler.ListCleaningServices)
		api.GET("/services/repair", repairHandler.ListRepairServices)
		api.GET("/services/health", healthHandler.ListHealthServices)
		api.GET("/services/b2b", b2bHandler.ListB2BServices)
		api.GET("/services/delivery", deliveryHandler.ListDeliveryOptions)
		
		// Travel & Transport
		api.GET("/bus/routes", busHandler.ListBusRoutes)
		api.GET("/flights", flightsHandler.ListFlights)
		api.GET("/tours", travelHandler.ListTours)
		api.GET("/drivers/nearby", taxiHandler.GetNearbyDrivers)
		
		// Accomodation
		api.GET("/properties", hostHandler.ListProperties)
		api.GET("/properties/:id", hostHandler.GetProperty)
		
		// Entertainment & Jobs
		api.GET("/cinema/movies/now-playing", cinemaHandler.ListNowPlayingMovies)
		api.GET("/cinema/movies/coming-soon", cinemaHandler.ListComingSoonMovies)
		api.GET("/cinema/movies/:id", cinemaHandler.GetMovieDetails)
		api.GET("/cinema/refreshments", cinemaHandler.ListRefreshments)
		api.GET("/cinema/movies/:id/showtimes", cinemaHandler.ListMovieShowtimes)
		api.GET("/jobs", jobsHandler.ListJobs)
		
		// C2C Marketplace
		api.GET("/c2c/listings", c2cHandler.ListC2CListings)
		api.GET("/c2c/listings/:id", c2cHandler.GetC2CListing)
		
		// Common
		api.GET("/businesses", businessHandler.ListBusinesses)
		api.GET("/businesses/:id", businessHandler.GetBusinessProfile)
		api.GET("/reviews", reviewHandler.GetReviewsByTarget)

		// --- Protected RBAC routes ---
		authRequired := api.Group("/")
		authRequired.Use(RequireAuth())
		{
			authRequired.GET("/users/me", func(c *gin.Context) {
				userID, _ := c.Get("user_id")
				uid := userID.(string)

				var user struct {
					ID          string `json:"id"`
					Email       string `json:"email"`
					FirstName   string `json:"first_name"`
					LastName    string `json:"last_name"`
					ProfileUrl  string `json:"profile_picture_url"`
				}

				err := conn.QueryRowContext(c.Request.Context(),
					"SELECT id, email, first_name, COALESCE(last_name, ''), COALESCE(profile_picture_url, '') FROM users WHERE id = $1",
					uid).Scan(&user.ID, &user.Email, &user.FirstName, &user.LastName, &user.ProfileUrl)

				if err != nil {
					c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to map identity"})
					return
				}

				rolesStr, _ := c.Get("user_roles")
				roles := strings.Split(rolesStr.(string), ",")

				c.JSON(http.StatusOK, gin.H{
					"user":  user,
					"roles": roles,
				})
			})

			// Business Onboarding
			authRequired.POST("/businesses", businessHandler.RegisterBusiness)
			authRequired.GET("/businesses/me", businessHandler.GetMyMall)

			// Product Management
			authRequired.POST("/products", ecommerceHandler.CreateProduct)
			authRequired.PUT("/products/:id", ecommerceHandler.UpdateProduct)
			authRequired.DELETE("/products/:id", ecommerceHandler.DeleteProduct)

			// Category Management
			authRequired.POST("/categories", ecommerceHandler.CreateCategory)
			authRequired.PUT("/categories/:id", ecommerceHandler.UpdateCategory)
			authRequired.DELETE("/categories/:id", ecommerceHandler.DeleteCategory)

			// Brand Management
			authRequired.POST("/brands", ecommerceHandler.CreateBrand)
			authRequired.PUT("/brands/:id", ecommerceHandler.UpdateBrand)
			authRequired.DELETE("/brands/:id", ecommerceHandler.DeleteBrand)

			// Model Management
			authRequired.POST("/models", ecommerceHandler.CreateProductModel)
			authRequired.PUT("/models/:id", ecommerceHandler.UpdateProductModel)
			authRequired.DELETE("/models/:id", ecommerceHandler.DeleteProductModel)

			// Cart & Checkout
			authRequired.GET("/cart", cartHandler.GetCart)
			authRequired.POST("/cart", cartHandler.AddToCart)
			authRequired.PUT("/cart/:id", ecommerceHandler.UpdateCartItemQuantity)
			authRequired.DELETE("/cart/:id", cartHandler.RemoveCartItem)
			authRequired.POST("/checkout", ecommerceHandler.Checkout)
			authRequired.GET("/orders", ecommerceHandler.GetOrders)

			// Service Bookings
			authRequired.POST("/services/book", serviceHandler.BookService)
			authRequired.GET("/services/my-bookings", serviceHandler.GetMyBookings)
			authRequired.POST("/services/bookings/:id/status", serviceHandler.ProviderUpdateBookingStatus)
			authRequired.POST("/services/listings", serviceHandler.CreateServiceListing)

			// Taxi Operations
			authRequired.POST("/taxi/location", taxiHandler.UpdateLocation)
			authRequired.POST("/taxi/status", taxiHandler.UpdateStatus)
			authRequired.POST("/taxi/request", taxiHandler.RequestRide)
			authRequired.POST("/food/delivery/assign", foodHandler.AssignDelivery)

			// Property Bookings
			authRequired.POST("/properties/book", propertyHandler.BookProperty)
			authRequired.POST("/properties/listings", propertyHandler.CreatePropertyListing)

			// C2C Actions
			authRequired.POST("/c2c/listings", c2cHandler.CreateC2CListing)

			// Review Management
			authRequired.POST("/reviews", reviewHandler.CreateReview)

			// Finance & Utilities
			authRequired.GET("/wallet/balance", payHandler.GetWalletBalance)
			authRequired.GET("/bills", billsHandler.ListUserBills)
			
			// B2B Marketplace
			authRequired.GET("/b2b/dashboard", b2bHandler.GetB2BDashboard)
			authRequired.GET("/b2b/items", b2bHandler.ListWholesaleItems)
			authRequired.GET("/b2b/items/:id", b2bHandler.GetWholesaleItem)
			authRequired.POST("/b2b/rfqs", b2bHandler.CreateRFQ)
			authRequired.GET("/b2b/rfqs", b2bHandler.ListMyRFQs)
			authRequired.POST("/b2b/quotes", b2bHandler.SubmitQuote)

			// Communication
			authRequired.GET("/messages", userHandler.ListMessages)
			authRequired.GET("/notifications", userHandler.ListNotifications)
		}
	}

	// Admin API with Swagger Documentation
	adminApi := r.Group("/admin")
	adminApi.Use(AuthMiddleware(authClient, conn))
	adminApi.Use(middleware.RequireRole("executive_admin", "admin"))
	{
		// Serve Swagger docs securely
		adminApi.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	}

	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}
	r.Run(":" + port)
}